Modif Dao Reservation

+++ src/main/java/tpVol/dao/DaoReservation.java
@@ -2,6 +2,6 @@ package tpVol.dao;
 
 import tpVol.model.Reservation;
 
-public interface DaoReservation extends DaoGeneric<Reservation, Integer> {
+public interface DaoReservation extends DaoGeneric<Reservation, Long> {
 
 }
